Carbon-Footprint-Management-System/backend/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
Werkzeug==2.0.1
pytesseract==0.3.8
Pillow==8.3.2
pandas==1.3.3
requests==2.26.0
fpdf==1.7.2
matplotlib==3.4.3 